+++ oxlint.config.ts
@@ -3,4 +3,7 @@ import { defineConfig } from 'oxlint';
 
 export default defineConfig({
   extends: [typescript, config, vitest],
+  env: {
+    browser: true,
+  },
 });