+++ src/header/WithHeaders.ts
@@ -1,0 +1,1 @@
+export type WithHeaders = Pick<Response, 'headers'>;